creando el safere con estilos

+++ src/Scripts/EstudiantesRegistro.js
@@ -8,10 +8,16 @@ loginRegistro.addEventListener('submit', (e) => {
 
     const Users = JSON.parse(localStorage.getItem('users')) || []
     const isUserRegistered = Users.find(user => user.name === name)
+    const confirmPassword = document.querySelector('#confirmPassword').value
+
+    if (confirmPassword !== password) {
+        return alert('Las Contraseñas no coinciden')
+    }
+
     if (isUserRegistered) {
         return alert('El nombre de usuario ya existe')
     }
-    Users.push({ name: name, password: password })
+    Users.push({ name: name, password: password, confirmPassword: confirmPassword })
     localStorage.setItem('users', JSON.stringify(Users))
     alert('Registro Existoso!')
     window.location.href = 'EstudiantesIngresar.html'
