+++ src/Scripts/ProfesoresRegistro.js
@@ -5,13 +5,17 @@ loginRegistro.addEventListener('submit', (e) => {
 
     const name = document.querySelector('#name').value
     const password = document.querySelector('#password').value
+    const confirmPassword = document.querySelector('#confirmPassword').value
 
     const Users = JSON.parse(localStorage.getItem('users')) || []
     const isUserRegistered = Users.find(user => user.name === name)
     if (isUserRegistered) {
         return alert('El nombre de usuario ya existe')
     }
-    Users.push({ name: name, password: password })
+    if (confirmPassword !== password) {
+        return alert('Las Contraseñas no coinciden')
+    }
+    Users.push({ name: name, password: password, confirmPassword: confirmPassword })
     localStorage.setItem('users', JSON.stringify(Users))
     alert('Registro Existoso!')
     window.location.href = 'ProfesoresIngresar.html'
